Create leds.py
This python script uses three optional LEDs on the Twisted Pear BBB/Hermes board to display operating parameters from the BBB.  When used to drive a RGB LED it produces color codes for the different sample rates, etc.  It works better with MPD than with Squeezelite, which does not update during pauses.