Create netio_server.py
This script is an adaptation of the RPi demo script from Netio.  It creates a TCP connection to the BBB and supports control from and data return to an iPad system controller.  NOTE: the code not secure!  I do not and would not enable wifi access to ports 8192 and 8193.  Additional work will be required to optimize this script and the subprocess calls it contains.